requirements.txt
annotated-types==0.5.0
anyio==4.0.0
appdirs==1.4.4
black==23.9.1
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
flake8==6.1.0
h11==0.14.0
httptools==0.6.0
idna==3.4
mccabe==0.7.0
motor==3.3.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
pip-review==1.3.0
platformdirs==3.10.0
pycodestyle==2.11.0
pydantic==2.3.0
pydantic_core==2.10.0
pyflakes==3.1.0
pymongo==4.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
sniffio==1.3.0
starlette==0.31.1
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.5
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
